re-organized admin tools and fixed types
