+++ aksel.nav.no/website/sanity/config.ts
@@ -1,8 +1,10 @@
+import { ClientConfig } from "@sanity/client";
+
 export const SANITY_PROJECT_ID = "hnbe3yhs";
-export const SANITY_API_VERSION = "2021-10-21";
+export const SANITY_API_VERSION = "2024-04-11";
 export const SANITY_DATASET = "production";
 
-export const clientConfig = {
+export const clientConfig: ClientConfig = {
   projectId: SANITY_PROJECT_ID,
   dataset: SANITY_DATASET,
   useCdn: false,
