+++ aksel.nav.no/website/sanity/schema/documents/admin/redirects.ts
